operator 부분을 별도의 클래스로 분리했습니다.
